Poprawa działania analityki oraz poprawa wyszukiwania najlepszej wartości + branie pod uwagę ograniczeń Problemu Optymalizacyjnego przy tym poszukiwaniu

+++ AlgorytmyDoTTP/AlgorytmyDoTTP/Struktura/Algorytmy/Losowy/RS.cs
@@ -42,7 +42,7 @@ namespace AlgorytmyDoTTP.Struktura.Algorytmy.Losowy
                     while (analityka.IleCzasuDzialaAlgorytm("s") < analityka.ZwrocCzasDzialaniaAlgorytmu())
                     {
                         losowanie.SzukajNajlepszegoRozwiazania(iloscRozwiazan, iloscElementow);
-                        analityka.DopiszWartoscProcesu(i, (short)analityka.IleCzasuDzialaAlgorytm("s"), losowanie.ZwrocNajlepszeRozwiazanie());
+                        analityka.DopiszWartoscProcesu(i, (int)analityka.IleCzasuDzialaAlgorytm("s"), losowanie.ZwrocNajlepszeRozwiazanie());
 
                         if (poprzedniaSekunda == -1 || poprzedniaSekunda != (int)analityka.IleCzasuDzialaAlgorytm("s"))
                         {
